Use rank instead of size.
Instead of keeping track of the size of each set, and putting the
smaller underneath the larger, we will keep track of the rank of
a set. New sets start with a rank of 0. If we unite two sets, we
put the set with the lesser rank under the set with the larger set.
Ranks will not change. If both sets have the same rank, we put one
set under the other, and increment the rank of the resulting set by 1.

This will improve the running time from (roughly) O (n log n) to
O (n alpha(n)), where alpha(n) is the inverse Ackermann function.